+++ tools/query_data/modules/chain_config.py
@@ -7,11 +7,14 @@ from langchain_core.output_parsers import StrOutputParser
 from langchain.globals import set_verbose
 
 from modules import metadata, api
-from agents import pandas
 
 # defines the langtab chain
 def create_chain():
-    set_verbose(True)
+    # Check if the environment variable ENV is set to 'development'
+    if os.getenv('ENV') == 'development':
+        set_verbose(True)
+    else:
+        set_verbose(False)
 
     # 1. Prompt template incorporating datasource metadata
     headless_bi_prompt_string = metadata.instantiate_prompt()
@@ -27,16 +30,10 @@ def create_chain():
     # 3. Standard Langchain parser
     output_parser = StrOutputParser()
 
-    # 4. Langchain pandas agent
-    pandas_agent = pandas.analyze
-
-    # 5. natural language response in the expected API format
+    # 4. natural language response in the expected API format
     json_parser = api.vds_transform
 
     # this chain defines the flow of data through the system
-
-    # chain = active_prompt_template | llm | output_parser | pandas_agent | json_parser
-
     chain = active_prompt_template | llm | output_parser | json_parser
