+++ tools/query_data/modules/prompts/nlq_to_vds.py
@@ -1,4 +1,4 @@
-from prompts import vds_schema, few_shot
+from modules.prompts import vds_schema, few_shot
 
 
 instructions = """
